pyproject.toml
[tool.poetry]
name = "tezca"
version = "0.1.0"
description = "Mexican Open Law Engine"
authors = ["madfam-org"]
license = "AGPL-3.0"
readme = "README.md"
packages = [{include = "apps"}, {include = "engines"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0"
djangorestframework = "^3.14"
django-cors-headers = "^4.3.1"
celery = "^5.3"
redis = "^5.0"
# XML Processing
lxml = "^5.1"
# Search
elasticsearch = "^8.0"
# NLP
spacy = "^3.7"
# Scraping
juriscraper = "^2.5"
beautifulsoup4 = "^4.12"
requests = "^2.31"
# Simulation
# openfisca-core = "*" (Blocked on Py3.11/Numpy)
# numpy = "^1.26"
drf-spectacular = "^0.29.0"
django-celery-beat = "^2.8.1"
pdfplumber = "^0.11.9"
psycopg2-binary = "^2.9.11"
gunicorn = "^23.0"
PyJWT = {version = "^2.8", extras = ["crypto"]}
cryptography = "^44.0"
# Sentry error tracking
sentry-sdk = {version = "^2.0", extras = ["django", "celery"], optional = true}
# Cloud storage (Cloudflare R2 / S3-compatible)
boto3 = {version = "^1.34", optional = true}
# PDF export
weasyprint = {version = "^62.0", optional = true}
# Extended export formats
python-docx = {version = "^1.1", optional = true}
ebooklib = {version = "^0.18", optional = true}
jinja2 = {version = "^3.1", optional = true}
# OCR pipeline (requires system tesseract-ocr + tesseract-ocr-spa)
pytesseract = {version = "^0.3.10", optional = true}
pdf2image = {version = "^1.16", optional = true}
# Excel parsing for QR state scraper
openpyxl = {version = "^3.1", optional = true}

[tool.poetry.extras]
r2 = ["boto3"]
sentry = ["sentry-sdk"]
pdf = ["weasyprint"]
export = ["weasyprint", "python-docx", "ebooklib", "jinja2"]
ocr = ["pytesseract", "pdf2image"]
scraping = ["openpyxl"]
production = ["boto3", "sentry-sdk", "weasyprint", "python-docx", "ebooklib", "jinja2", "pytesseract", "pdf2image", "openpyxl"]

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
pytest-cov = "^4.1"
pytest-django = "^4.7"
black = "^24.1"
isort = "^5.13"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "apps.indigo.settings"
pythonpath = ["."]

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
